fix(harvest): share effective remote path resolution
